server/requirements.txt
requests>=2.31.0
Pillow>=10.0.0
python-bidi>=0.4.2
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
python-dateutil>=2.8.2
pytz>=2023.3
